Move some datetime functions to @openshift-console/plugin-shared
Move formatPrometheusDuration and parsePrometheusDuration to
@openshift-console/plugin-shared
