+++ frontend/packages/console-plugin-shared/src/datetime/index.ts
@@ -1,0 +1,1 @@
+export * from './prometheus';
