+++ frontend/packages/dev-console/src/components/monitoring/alerts/SilenceDurationDropdown.tsx
@@ -1,4 +1,5 @@
 import * as React from 'react';
+import { parsePrometheusDuration } from '@openshift-console/plugin-shared/src/datetime/prometheus';
 import * as _ from 'lodash';
 import { useTranslation } from 'react-i18next';
 // FIXME upgrading redux types is causing many errors at this time
@@ -18,7 +19,6 @@ import { coFetchJSON } from '@console/internal/co-fetch';
 import { ALERTMANAGER_TENANCY_BASE_PATH } from '@console/internal/components/graphs';
 import { isSilenced } from '@console/internal/components/monitoring/utils';
 import { Dropdown, LoadingInline } from '@console/internal/components/utils';
-import { parsePrometheusDuration } from '@console/internal/components/utils/datetime';
 import { RootState } from '@console/internal/redux';
 import './SilenceDurationDropdown.scss';
