+++ frontend/packages/pipelines-plugin/src/components/pipelines/pipeline-metrics/PipelineRunDurationGraph.tsx
@@ -1,11 +1,11 @@
 import * as React from 'react';
+import { formatPrometheusDuration } from '@openshift-console/plugin-shared/src/datetime/prometheus';
 import { ChartThemeColor, ChartVoronoiContainer } from '@patternfly/react-charts';
 import { Bullseye, Flex, FlexItem, Grid, GridItem } from '@patternfly/react-core';
 import { useTranslation } from 'react-i18next';
 import Measure from 'react-measure';
 import { GraphEmpty } from '@console/internal/components/graphs/graph-empty';
 import { LoadingInline, truncateMiddle } from '@console/internal/components/utils';
-import { formatPrometheusDuration } from '@console/internal/components/utils/datetime';
 import { DEFAULT_LEGEND_CHART_HEIGHT } from '../const';
 import { usePipelineRunDurationPoll } from '../hooks';
 import { LineChart } from './charts/lineChart';
