+++ frontend/public/components/utils/datetime.ts
@@ -136,46 +136,6 @@ export const fromNow = (dateTime: string | Date, now?: Date, options?) => {
 
 export const isValid = (dateTime: Date) => dateTime instanceof Date && !_.isNaN(dateTime.valueOf());
 
-// Conversions between units and milliseconds
-const s = 1000;
-const m = s * 60;
-const h = m * 60;
-const d = h * 24;
-const w = d * 7;
-const units = { w, d, h, m, s };
-
-// Formats a duration in milliseconds like "1h 10m"
-export const formatPrometheusDuration = (ms: number) => {
-  if (!_.isFinite(ms) || ms < 0) {
-    return '';
-  }
-  let remaining = ms;
-  let str = '';
-  _.each(units, (factor, unit) => {
-    const n = Math.floor(remaining / factor);
-    if (n > 0) {
-      str += `${n}${unit} `;
-      remaining -= n * factor;
-    }
-  });
-  return _.trim(str);
-};
-
-// Converts a duration like "1h 10m 23s" to milliseconds or returns 0 if the duration could not be
-// parsed
-export const parsePrometheusDuration = (duration: string): number => {
-  try {
-    const parts = duration
-      .trim()
-      .split(/\s+/)
-      .map((p) => p.match(/^(\d+)([wdhms])$/));
-    return _.sumBy(parts, (p) => parseInt(p[1], 10) * units[p[2]]);
-  } catch (ignored) {
-    // Invalid duration format
-    return 0;
-  }
-};
-
 const zeroPad = (number: number) => (number < 10 ? `0${number}` : number);
 
 export const twentyFourHourTime = (date: Date, showSeconds?: boolean): string => {
